Begin moving OO solvers out of model files
Had to move one very small agent type into its own file. It looks like it's a PortfolioConsumerType with a slightly different solver.
