+++ HARK/ConsumptionSaving/ConsPortfolioModel.py
@@ -8,15 +8,10 @@ from copy import deepcopy
 
 import numpy as np
 
-from HARK import AgentType, NullFunc, make_one_period_oo_solver
+from HARK import AgentType, NullFunc
 from HARK.ConsumptionSaving.ConsIndShockModel import (
     IndShockConsumerType,
     init_idiosyncratic_shocks,
-    utility,
-    utility_inv,
-    utility_invP,
-    utilityP,
-    utilityP_inv,
 )
 from HARK.ConsumptionSaving.ConsRiskyAssetModel import RiskyAssetConsumerType
 from HARK.distribution import expected
@@ -310,19 +305,6 @@ class PortfolioConsumerType(RiskyAssetConsumerType):
         # Store controls as attributes of self
         self.controls["cNrm"] = cNrmNow
         self.controls["Share"] = ShareNow
-
-
-class SequentialPortfolioConsumerType(PortfolioConsumerType):
-    def __init__(self, verbose=False, quiet=False, **kwds):
-        params = init_portfolio.copy()
-        params.update(kwds)
-        kwds = params
-
-        # Initialize a basic consumer type
-        PortfolioConsumerType.__init__(self, verbose=verbose, quiet=quiet, **kwds)
-
-        # Set the solver for the portfolio model, and update various constructed attributes
-        self.solve_one_period = make_one_period_oo_solver(ConsSequentialPortfolioSolver)
 
 
 def solve_one_period_ConsPortfolio(
@@ -936,932 +918,6 @@ def solve_one_period_ConsPortfolio(
         EndOfPrddvds_fxd=save_points["eop_dvds_fxd"],
     )
     return solution_now
-
-
-class ConsPortfolioSolver(MetricObject):
-    """
-    Define an object-oriented one period solver.
-    Solve the one period problem for a portfolio-choice consumer.
-    This solver is used when the income and risky return shocks
-    are independent and the allowed optimal share is continuous.
-
-    Parameters
-    ----------
-    solution_next : PortfolioSolution
-        Solution to next period's problem.
-    ShockDstn : [np.array]
-        List with four arrays: discrete probabilities, permanent income shocks,
-        transitory income shocks, and risky returns.  This is only used if the
-        input IndepDstnBool is False, indicating that income and return distributions
-        can't be assumed to be independent.
-    IncShkDstn : distribution.Distribution
-        Discrete distribution of permanent income shocks
-        and transitory income shocks.  This is only used if the input IndepDsntBool
-        is True, indicating that income and return distributions are independent.
-    RiskyDstn : [np.array]
-        List with two arrays: discrete probabilities and risky asset returns. This
-        is only used if the input IndepDstnBool is True, indicating that income
-        and return distributions are independent.
-    LivPrb : float
-        Survival probability; likelihood of being alive at the beginning of
-        the succeeding period.
-    DiscFac : float
-        Intertemporal discount factor for future utility.
-    CRRA : float
-        Coefficient of relative risk aversion.
-    Rfree : float
-        Risk free interest factor on end-of-period assets.
-    PermGroFac : float
-        Expected permanent income growth factor at the end of this period.
-    BoroCnstArt: float or None
-        Borrowing constraint for the minimum allowable assets to end the
-        period with.  In this model, it is *required* to be zero.
-    aXtraGrid: np.array
-        Array of "extra" end-of-period asset values-- assets above the
-        absolute minimum acceptable level.
-    ShareGrid : np.array
-        Array of risky portfolio shares on which to define the interpolation
-        of the consumption function when Share is fixed.
-    vFuncBool: boolean
-        An indicator for whether the value function should be computed and
-        included in the reported solution.
-    AdjustPrb : float
-        Probability that the agent will be able to update his portfolio share.
-    DiscreteShareBool : bool
-        Indicator for whether risky portfolio share should be optimized on the
-        continuous [0,1] interval using the FOC (False), or instead only selected
-        from the discrete set of values in ShareGrid (True).  If True, then
-        vFuncBool must also be True.
-    ShareLimit : float
-        Limiting lower bound of risky portfolio share as mNrm approaches infinity.
-    IndepDstnBool : bool
-        Indicator for whether the income and risky return distributions are in-
-        dependent of each other, which can speed up the expectations step.
-    """
-
-    def __init__(
-        self,
-        solution_next,
-        ShockDstn,
-        IncShkDstn,
-        RiskyDstn,
-        LivPrb,
-        DiscFac,
-        CRRA,
-        Rfree,
-        PermGroFac,
-        BoroCnstArt,
-        aXtraGrid,
-        ShareGrid,
-        vFuncBool,
-        AdjustPrb,
-        DiscreteShareBool,
-        ShareLimit,
-        IndepDstnBool,
-    ):
-        """
-        Constructor for portfolio choice problem solver.
-        """
-
-        self.solution_next = solution_next
-        self.ShockDstn = ShockDstn
-        self.IncShkDstn = IncShkDstn
-        self.RiskyDstn = RiskyDstn
-        self.LivPrb = LivPrb
-        self.DiscFac = DiscFac
-        self.CRRA = CRRA
-        self.Rfree = Rfree
-        self.PermGroFac = PermGroFac
-        self.BoroCnstArt = BoroCnstArt
-        self.aXtraGrid = aXtraGrid
-        self.ShareGrid = ShareGrid
-        self.vFuncBool = vFuncBool
-        self.AdjustPrb = AdjustPrb
-        self.DiscreteShareBool = DiscreteShareBool
-        self.ShareLimit = ShareLimit
-        self.IndepDstnBool = IndepDstnBool
-
-        # Make sure the individual is liquidity constrained.  Allowing a consumer to
-        # borrow *and* invest in an asset with unbounded (negative) returns is a bad mix.
-        if BoroCnstArt != 0.0:
-            raise ValueError("PortfolioConsumerType must have BoroCnstArt=0.0!")
-
-        # Make sure that if risky portfolio share is optimized only discretely, then
-        # the value function is also constructed (else this task would be impossible).
-        if DiscreteShareBool and (not vFuncBool):
-            raise ValueError(
-                "PortfolioConsumerType requires vFuncBool to be True when DiscreteShareBool is True!"
-            )
-
-        self.def_utility_funcs()
-
-    def def_utility_funcs(self):
-        """
-        Define temporary functions for utility and its derivative and inverse
-        """
-
-        self.u = lambda x: utility(x, self.CRRA)
-        self.uP = lambda x: utilityP(x, self.CRRA)
-        self.uPinv = lambda x: utilityP_inv(x, self.CRRA)
-        self.uinv = lambda x: utility_inv(x, self.CRRA)
-        self.uinvP = lambda x: utility_invP(x, self.CRRA)
-
-    def set_and_update_values(self):
-        """
-        Unpacks some of the inputs (and calculates simple objects based on them),
-        storing the results in self for use by other methods.
-        """
-
-        # Unpack next period's solution
-        self.vPfuncAdj_next = self.solution_next.vPfuncAdj
-        self.dvdmFuncFxd_next = self.solution_next.dvdmFuncFxd
-        self.dvdsFuncFxd_next = self.solution_next.dvdsFuncFxd
-        self.vFuncAdj_next = self.solution_next.vFuncAdj
-        self.vFuncFxd_next = self.solution_next.vFuncFxd
-
-        # Unpack the shock distribution
-        TranShks_next = self.IncShkDstn.atoms[1]
-
-        # Flag for whether the natural borrowing constraint is zero
-        self.zero_bound = np.min(TranShks_next) == 0.0
-
-    def prepare_to_solve(self):
-        """
-        Perform preparatory work.
-        """
-
-        self.set_and_update_values()
-
-    def prepare_to_calc_EndOfPrdvP(self):
-        """
-        Prepare to calculate end-of-period marginal values by creating an array
-        of market resources that the agent could have next period, considering
-        the grid of end-of-period assets and the distribution of shocks he might
-        experience next period.
-        """
-
-        # Unpack the shock distribution
-        Risky_next = self.RiskyDstn.atoms
-        RiskyMax = np.max(Risky_next)
-        RiskyMin = np.min(Risky_next)
-
-        # bNrm represents R*a, balances after asset return shocks but before income.
-        # This just uses the highest risky return as a rough shifter for the aXtraGrid.
-        if self.zero_bound:
-            self.aNrmGrid = self.aXtraGrid
-            self.bNrmGrid = np.insert(
-                RiskyMax * self.aXtraGrid, 0, RiskyMin * self.aXtraGrid[0]
-            )
-        else:
-            # Add an asset point at exactly zero
-            self.aNrmGrid = np.insert(self.aXtraGrid, 0, 0.0)
-            self.bNrmGrid = RiskyMax * np.insert(self.aXtraGrid, 0, 0.0)
-
-        # Get grid and shock sizes, for easier indexing
-        self.aNrmCount = self.aNrmGrid.size
-        self.ShareCount = self.ShareGrid.size
-
-        # Make tiled arrays to calculate future realizations of mNrm and Share when integrating over IncShkDstn
-        self.bNrmNext, self.ShareNext = np.meshgrid(
-            self.bNrmGrid, self.ShareGrid, indexing="ij"
-        )
-
-    def m_nrm_next(self, shocks, b_nrm_next):
-        """
-        Calculate future realizations of market resources
-        """
-
-        return b_nrm_next / (shocks["PermShk"] * self.PermGroFac) + shocks["TranShk"]
-
-    def calc_EndOfPrdvP(self):
-        """
-        Calculate end-of-period marginal value of assets and shares at each point
-        in aNrm and ShareGrid. Does so by taking expectation of next period marginal
-        values across income and risky return shocks.
-        """
-
-        def dvdb_dist(shocks, b_nrm, Share_next):
-            """
-            Evaluate realizations of marginal value of market resources next period
-            """
-
-            mNrm_next = self.m_nrm_next(shocks, b_nrm)
-
-            dvdmAdj_next = self.vPfuncAdj_next(mNrm_next)
-            if self.AdjustPrb < 1.0:
-                # Expand to the same dimensions as mNrm
-                Share_next_expanded = Share_next + np.zeros_like(mNrm_next)
-                dvdmFxd_next = self.dvdmFuncFxd_next(mNrm_next, Share_next_expanded)
-                # Combine by adjustment probability
-                dvdm_next = (
-                    self.AdjustPrb * dvdmAdj_next
-                    + (1.0 - self.AdjustPrb) * dvdmFxd_next
-                )
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                dvdm_next = dvdmAdj_next
-
-            return (shocks["PermShk"] * self.PermGroFac) ** (-self.CRRA) * dvdm_next
-
-        def dvds_dist(shocks, b_nrm, Share_next):
-            """
-            Evaluate realizations of marginal value of risky share next period
-            """
-
-            mNrm_next = self.m_nrm_next(shocks, b_nrm)
-            # No marginal value of Share if it's a free choice!
-            dvdsAdj_next = np.zeros_like(mNrm_next)
-            if self.AdjustPrb < 1.0:
-                # Expand to the same dimensions as mNrm
-                Share_next_expanded = Share_next + np.zeros_like(mNrm_next)
-                dvdsFxd_next = self.dvdsFuncFxd_next(mNrm_next, Share_next_expanded)
-                # Combine by adjustment probability
-                dvds_next = (
-                    self.AdjustPrb * dvdsAdj_next
-                    + (1.0 - self.AdjustPrb) * dvdsFxd_next
-                )
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                dvds_next = dvdsAdj_next
-
-            return (shocks["PermShk"] * self.PermGroFac) ** (
-                1.0 - self.CRRA
-            ) * dvds_next
-
-        # Calculate intermediate marginal value of bank balances by taking expectations over income shocks
-        dvdb_intermed = self.IncShkDstn.expected(
-            dvdb_dist, self.bNrmNext, self.ShareNext
-        )
-
-        dvdbNvrs_intermed = self.uPinv(dvdb_intermed)
-        dvdbNvrsFunc_intermed = BilinearInterp(
-            dvdbNvrs_intermed, self.bNrmGrid, self.ShareGrid
-        )
-        dvdbFunc_intermed = MargValueFuncCRRA(dvdbNvrsFunc_intermed, self.CRRA)
-
-        # Calculate intermediate marginal value of risky portfolio share by taking expectations
-        dvds_intermed = self.IncShkDstn.expected(
-            dvds_dist, self.bNrmNext, self.ShareNext
-        )
-
-        dvdsFunc_intermed = BilinearInterp(dvds_intermed, self.bNrmGrid, self.ShareGrid)
-
-        # Make tiled arrays to calculate future realizations of bNrm and Share when integrating over RiskyDstn
-        self.aNrm_tiled, self.ShareNext = np.meshgrid(
-            self.aNrmGrid, self.ShareGrid, indexing="ij"
-        )
-
-        # Evaluate realizations of value and marginal value after asset returns are realized
-
-        def EndOfPrddvda_dist(shock, a_nrm, Share_next):
-            # Calculate future realizations of bank balances bNrm
-            Rxs = shock - self.Rfree
-            Rport = self.Rfree + Share_next * Rxs
-            b_nrm_next = Rport * a_nrm
-
-            # Ensure shape concordance
-            Share_next_rep = Share_next + np.zeros_like(b_nrm_next)
-
-            return Rport * dvdbFunc_intermed(b_nrm_next, Share_next_rep)
-
-        def EndOfPrddvds_dist(shock, a_nrm, Share_next):
-            # Calculate future realizations of bank balances bNrm
-            Rxs = shock - self.Rfree
-            Rport = self.Rfree + Share_next * Rxs
-            b_nrm_next = Rport * a_nrm
-
-            # Make the shares match the dimension of b, so that it can be vectorized
-            Share_next_expand = Share_next + np.zeros_like(b_nrm_next)
-
-            return Rxs * a_nrm * dvdbFunc_intermed(
-                b_nrm_next, Share_next_expand
-            ) + dvdsFunc_intermed(b_nrm_next, Share_next_expand)
-
-        # Calculate end-of-period marginal value of assets by taking expectations
-        self.EndOfPrddvda = (
-            self.DiscFac
-            * self.LivPrb
-            * self.RiskyDstn.expected(
-                EndOfPrddvda_dist, self.aNrm_tiled, self.ShareNext
-            )
-        )
-
-        self.EndOfPrddvdaNvrs = self.uPinv(self.EndOfPrddvda)
-
-        # Calculate end-of-period marginal value of risky portfolio share by taking expectations
-        self.EndOfPrddvds = (
-            self.DiscFac
-            * self.LivPrb
-            * self.RiskyDstn.expected(
-                EndOfPrddvds_dist, self.aNrm_tiled, self.ShareNext
-            )
-        )
-
-    def optimize_share(self):
-        """
-        Optimization of Share on continuous interval [0,1]
-        """
-
-        FOC_s = self.EndOfPrddvds
-
-        # For each value of aNrm, find the value of Share such that FOC-Share == 0.
-        crossing = np.logical_and(FOC_s[:, 1:] <= 0.0, FOC_s[:, :-1] >= 0.0)
-        share_idx = np.argmax(crossing, axis=1)
-        a_idx = np.arange(self.aNrmCount)
-        bot_s = self.ShareGrid[share_idx]
-        top_s = self.ShareGrid[share_idx + 1]
-        bot_f = FOC_s[a_idx, share_idx]
-        top_f = FOC_s[a_idx, share_idx + 1]
-        bot_c = self.EndOfPrddvdaNvrs[a_idx, share_idx]
-        top_c = self.EndOfPrddvdaNvrs[a_idx, share_idx + 1]
-        alpha = 1.0 - top_f / (top_f - bot_f)
-
-        self.Share_now = (1.0 - alpha) * bot_s + alpha * top_s
-        self.cNrmAdj_now = (1.0 - alpha) * bot_c + alpha * top_c
-
-        # If agent wants to put more than 100% into risky asset, he is constrained
-        constrained_top = FOC_s[:, -1] > 0.0
-        # Likewise if he wants to put less than 0% into risky asset
-        constrained_bot = FOC_s[:, 0] < 0.0
-
-        # For values of aNrm at which the agent wants to put
-        # more than 100% into risky asset, constrain them
-        self.Share_now[constrained_top] = 1.0
-        self.Share_now[constrained_bot] = 0.0
-
-        # Get consumption when share-constrained
-        self.cNrmAdj_now[constrained_top] = self.EndOfPrddvdaNvrs[constrained_top, -1]
-        self.cNrmAdj_now[constrained_bot] = self.EndOfPrddvdaNvrs[constrained_bot, 0]
-
-        if not self.zero_bound:
-            # aNrm=0, so there's no way to "optimize" the portfolio
-            self.Share_now[0] = 1.0
-            # Consumption when aNrm=0 does not depend on Share
-            self.cNrmAdj_now[0] = self.EndOfPrddvdaNvrs[0, -1]
-
-    def make_basic_solution(self):
-        """
-        Given end of period assets and end of period marginal values, construct
-        the basic solution for this period.
-        """
-
-        # Calculate the endogenous mNrm gridpoints when the agent adjusts his portfolio
-        self.mNrmAdj_now = self.aNrmGrid + self.cNrmAdj_now
-
-        # Construct the consumption function when the agent can adjust
-        cNrmAdj_now = np.insert(self.cNrmAdj_now, 0, 0.0)
-        self.cFuncAdj_now = LinearInterp(
-            np.insert(self.mNrmAdj_now, 0, 0.0), cNrmAdj_now
-        )
-
-        # Construct the marginal value (of mNrm) function when the agent can adjust
-        self.vPfuncAdj_now = MargValueFuncCRRA(self.cFuncAdj_now, self.CRRA)
-
-        # Construct the consumption function when the agent *can't* adjust the risky share, as well
-        # as the marginal value of Share function
-        cFuncFxd_by_Share = []
-        dvdsFuncFxd_by_Share = []
-        for j in range(self.ShareCount):
-            cNrmFxd_temp = self.EndOfPrddvdaNvrs[:, j]
-            mNrmFxd_temp = self.aNrmGrid + cNrmFxd_temp
-            cFuncFxd_by_Share.append(
-                LinearInterp(
-                    np.insert(mNrmFxd_temp, 0, 0.0), np.insert(cNrmFxd_temp, 0, 0.0)
-                )
-            )
-            dvdsFuncFxd_by_Share.append(
-                LinearInterp(
-                    np.insert(mNrmFxd_temp, 0, 0.0),
-                    np.insert(self.EndOfPrddvds[:, j], 0, self.EndOfPrddvds[0, j]),
-                )
-            )
-        self.cFuncFxd_now = LinearInterpOnInterp1D(cFuncFxd_by_Share, self.ShareGrid)
-        self.dvdsFuncFxd_now = LinearInterpOnInterp1D(
-            dvdsFuncFxd_by_Share, self.ShareGrid
-        )
-
-        # The share function when the agent can't adjust his portfolio is trivial
-        self.ShareFuncFxd_now = IdentityFunction(i_dim=1, n_dims=2)
-
-        # Construct the marginal value of mNrm function when the agent can't adjust his share
-        self.dvdmFuncFxd_now = MargValueFuncCRRA(self.cFuncFxd_now, self.CRRA)
-
-    def make_ShareFuncAdj(self):
-        """
-        Construct the risky share function when the agent can adjust
-        """
-
-        if self.zero_bound:
-            Share_lower_bound = self.ShareLimit
-        else:
-            Share_lower_bound = 1.0
-        Share_now = np.insert(self.Share_now, 0, Share_lower_bound)
-        self.ShareFuncAdj_now = LinearInterp(
-            np.insert(self.mNrmAdj_now, 0, 0.0),
-            Share_now,
-            intercept_limit=self.ShareLimit,
-            slope_limit=0.0,
-        )
-
-    def add_save_points(self):
-        # This is a point at which (a,c,share) have consistent length. Take the
-        # snapshot for storing the grid and values in the solution.
-        self.save_points = {
-            "a": deepcopy(self.aNrmGrid),
-            "eop_dvda_adj": self.uP(self.cNrmAdj_now),
-            "share_adj": deepcopy(self.Share_now),
-            "share_grid": deepcopy(self.ShareGrid),
-            "eop_dvda_fxd": self.uP(self.EndOfPrddvda),
-            "eop_dvds_fxd": self.EndOfPrddvds,
-        }
-
-    def add_vFunc(self):
-        """
-        Creates the value function for this period and adds it to the solution.
-        """
-
-        self.make_EndOfPrdvFunc()
-        self.make_vFunc()
-
-    def make_EndOfPrdvFunc(self):
-        """
-        Construct the end-of-period value function for this period, storing it
-        as an attribute of self for use by other methods.
-        """
-
-        def v_intermed_dist(shocks, b_nrm, Share_next):
-            mNrm_next = self.m_nrm_next(shocks, b_nrm)
-
-            vAdj_next = self.vFuncAdj_next(mNrm_next)
-            if self.AdjustPrb < 1.0:
-                vFxd_next = self.vFuncFxd_next(mNrm_next, Share_next)
-                # Combine by adjustment probability
-                v_next = self.AdjustPrb * vAdj_next + (1.0 - self.AdjustPrb) * vFxd_next
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                v_next = vAdj_next
-
-            return (shocks["PermShk"] * self.PermGroFac) ** (1.0 - self.CRRA) * v_next
-
-        # Calculate intermediate value by taking expectations over income shocks
-        v_intermed = self.IncShkDstn.expected(
-            v_intermed_dist, self.bNrmNext, self.ShareNext
-        )
-
-        vNvrs_intermed = self.uinv(v_intermed)
-        vNvrsFunc_intermed = BilinearInterp(
-            vNvrs_intermed, self.bNrmGrid, self.ShareGrid
-        )
-        vFunc_intermed = ValueFuncCRRA(vNvrsFunc_intermed, self.CRRA)
-
-        def EndOfPrdv_dist(shock, a_nrm, Share_next):
-            # Calculate future realizations of bank balances bNrm
-            Rxs = shock - self.Rfree
-            Rport = self.Rfree + Share_next * Rxs
-            b_nrm_next = Rport * a_nrm
-
-            # Make an extended share_next of the same dimension as b_nrm so
-            # that the function can be vectorized
-            Share_next_extended = Share_next + np.zeros_like(b_nrm_next)
-
-            return vFunc_intermed(b_nrm_next, Share_next_extended)
-
-        # Calculate end-of-period value by taking expectations
-        self.EndOfPrdv = (
-            self.DiscFac
-            * self.LivPrb
-            * self.RiskyDstn.expected(EndOfPrdv_dist, self.aNrm_tiled, self.ShareNext)
-        )
-
-        self.EndOfPrdvNvrs = self.uinv(self.EndOfPrdv)
-
-    def make_vFunc(self):
-        """
-        Creates the value functions for this period, defined over market
-        resources m when agent can adjust his portfolio, and over market
-        resources and fixed share when agent can not adjust his portfolio.
-        self must have the attribute EndOfPrdvFunc in order to execute.
-        """
-
-        # First, make an end-of-period value function over aNrm and Share
-        EndOfPrdvNvrsFunc = BilinearInterp(
-            self.EndOfPrdvNvrs, self.aNrmGrid, self.ShareGrid
-        )
-        EndOfPrdvFunc = ValueFuncCRRA(EndOfPrdvNvrsFunc, self.CRRA)
-
-        # Construct the value function when the agent can adjust his portfolio
-        mNrm_temp = self.aXtraGrid  # Just use aXtraGrid as our grid of mNrm values
-        cNrm_temp = self.cFuncAdj_now(mNrm_temp)
-        aNrm_temp = mNrm_temp - cNrm_temp
-        Share_temp = self.ShareFuncAdj_now(mNrm_temp)
-        v_temp = self.u(cNrm_temp) + EndOfPrdvFunc(aNrm_temp, Share_temp)
-        vNvrs_temp = self.uinv(v_temp)
-        vNvrsP_temp = self.uP(cNrm_temp) * self.uinvP(v_temp)
-        vNvrsFuncAdj = CubicInterp(
-            np.insert(mNrm_temp, 0, 0.0),  # x_list
-            np.insert(vNvrs_temp, 0, 0.0),  # f_list
-            np.insert(vNvrsP_temp, 0, vNvrsP_temp[0]),  # dfdx_list
-        )
-        # Re-curve the pseudo-inverse value function
-        self.vFuncAdj_now = ValueFuncCRRA(vNvrsFuncAdj, self.CRRA)
-
-        # Construct the value function when the agent *can't* adjust his portfolio
-        mNrm_temp, Share_temp = np.meshgrid(self.aXtraGrid, self.ShareGrid)
-        cNrm_temp = self.cFuncFxd_now(mNrm_temp, Share_temp)
-        aNrm_temp = mNrm_temp - cNrm_temp
-        v_temp = self.u(cNrm_temp) + EndOfPrdvFunc(aNrm_temp, Share_temp)
-        vNvrs_temp = self.uinv(v_temp)
-        vNvrsP_temp = self.uP(cNrm_temp) * self.uinvP(v_temp)
-        vNvrsFuncFxd_by_Share = []
-        for j in range(self.ShareCount):
-            vNvrsFuncFxd_by_Share.append(
-                CubicInterp(
-                    np.insert(mNrm_temp[:, 0], 0, 0.0),  # x_list
-                    np.insert(vNvrs_temp[:, j], 0, 0.0),  # f_list
-                    np.insert(vNvrsP_temp[:, j], 0, vNvrsP_temp[j, 0]),  # dfdx_list
-                )
-            )
-        vNvrsFuncFxd = LinearInterpOnInterp1D(vNvrsFuncFxd_by_Share, self.ShareGrid)
-        self.vFuncFxd_now = ValueFuncCRRA(vNvrsFuncFxd, self.CRRA)
-
-    def make_porfolio_solution(self):
-        self.solution = PortfolioSolution(
-            cFuncAdj=self.cFuncAdj_now,
-            ShareFuncAdj=self.ShareFuncAdj_now,
-            vPfuncAdj=self.vPfuncAdj_now,
-            vFuncAdj=self.vFuncAdj_now,
-            cFuncFxd=self.cFuncFxd_now,
-            ShareFuncFxd=self.ShareFuncFxd_now,
-            dvdmFuncFxd=self.dvdmFuncFxd_now,
-            dvdsFuncFxd=self.dvdsFuncFxd_now,
-            vFuncFxd=self.vFuncFxd_now,
-            aGrid=self.save_points["a"],
-            Share_adj=self.save_points["share_adj"],
-            EndOfPrddvda_adj=self.save_points["eop_dvda_adj"],
-            ShareGrid=self.save_points["share_grid"],
-            EndOfPrddvda_fxd=self.save_points["eop_dvda_fxd"],
-            EndOfPrddvds_fxd=self.save_points["eop_dvds_fxd"],
-            AdjPrb=self.AdjustPrb,
-        )
-
-    def solve(self):
-        """
-        Solve the one period problem for a portfolio-choice consumer.
-
-        Returns
-        -------
-        solution_now : PortfolioSolution
-        The solution to the single period consumption-saving with portfolio choice
-        problem.  Includes two consumption and risky share functions: one for when
-        the agent can adjust his portfolio share (Adj) and when he can't (Fxd).
-        """
-
-        # Make arrays of end-of-period assets and end-of-period marginal values
-        self.prepare_to_calc_EndOfPrdvP()
-        self.calc_EndOfPrdvP()
-
-        # Construct a basic solution for this period
-        self.optimize_share()
-        self.make_basic_solution()
-        self.make_ShareFuncAdj()
-
-        self.add_save_points()
-
-        # Add the value function if requested
-        if self.vFuncBool:
-            self.add_vFunc()
-        else:  # If vFuncBool is False, fill in dummy values
-            self.vFuncAdj_now = NullFunc()
-            self.vFuncFxd_now = NullFunc()
-
-        self.make_porfolio_solution()
-
-        return self.solution
-
-
-class ConsPortfolioDiscreteSolver(ConsPortfolioSolver):
-    """
-    Define an object-oriented one period solver.
-    Solve the one period problem for a portfolio-choice consumer.
-    This solver is used when the income and risky return shocks
-    are independent and the allowed optimal share is discrete
-    over a finite set of points in ShareGrid.
-    """
-
-    def optimize_share(self):
-        """
-        Optimization of Share on the discrete set ShareGrid
-        """
-
-        opt_idx = np.argmax(self.EndOfPrdv, axis=1)
-        # Best portfolio share is one with highest value
-        self.Share_now = self.ShareGrid[opt_idx]
-        # Take cNrm at that index as well
-        self.cNrmAdj_now = self.EndOfPrddvdaNvrs[np.arange(self.aNrmCount), opt_idx]
-        if not self.zero_bound:
-            # aNrm=0, so there's no way to "optimize" the portfolio
-            self.Share_now[0] = 1.0
-            # Consumption when aNrm=0 does not depend on Share
-            self.cNrmAdj_now[0] = self.EndOfPrddvdaNvrs[0, -1]
-
-    def make_ShareFuncAdj(self):
-        """
-        Construct the risky share function when the agent can adjust
-        """
-
-        mNrmAdj_mid = (self.mNrmAdj_now[1:] + self.mNrmAdj_now[:-1]) / 2
-        mNrmAdj_plus = mNrmAdj_mid * (1.0 + 1e-12)
-        mNrmAdj_comb = (np.transpose(np.vstack((mNrmAdj_mid, mNrmAdj_plus)))).flatten()
-        mNrmAdj_comb = np.append(np.insert(mNrmAdj_comb, 0, 0.0), self.mNrmAdj_now[-1])
-        Share_comb = (
-            np.transpose(np.vstack((self.Share_now, self.Share_now)))
-        ).flatten()
-        self.ShareFuncAdj_now = LinearInterp(mNrmAdj_comb, Share_comb)
-
-    def solve(self):
-        """
-        Solve the one period problem for a portfolio-choice consumer.
-
-        Returns
-        -------
-        solution_now : PortfolioSolution
-        The solution to the single period consumption-saving with portfolio choice
-        problem.  Includes two consumption and risky share functions: one for when
-        the agent can adjust his portfolio share (Adj) and when he can't (Fxd).
-        """
-
-        # Make arrays of end-of-period assets and end-of-period marginal value
-        self.prepare_to_calc_EndOfPrdvP()
-        self.calc_EndOfPrdvP()
-
-        # Construct a basic solution for this period
-        self.make_EndOfPrdvFunc()
-        self.optimize_share()
-        self.make_basic_solution()
-        self.make_ShareFuncAdj()
-
-        self.add_save_points()
-
-        self.make_vFunc()
-
-        self.make_porfolio_solution()
-
-        return self.solution
-
-
-class ConsPortfolioJointDistSolver(ConsPortfolioDiscreteSolver, ConsPortfolioSolver):
-    """
-    Define an object-oriented one period solver.
-    Solve the one period problem for a portfolio-choice consumer.
-    This solver is used when the income and risky return shocks
-    are not independent. The optimal share can be continuous or
-    discrete.
-    """
-
-    def set_and_update_values(self):
-        """
-        Unpacks some of the inputs (and calculates simple objects based on them),
-        storing the results in self for use by other methods.
-        """
-
-        # Unpack next period's solution
-        self.vPfuncAdj_next = self.solution_next.vPfuncAdj
-        self.dvdmFuncFxd_next = self.solution_next.dvdmFuncFxd
-        self.dvdsFuncFxd_next = self.solution_next.dvdsFuncFxd
-        self.vFuncAdj_next = self.solution_next.vFuncAdj
-        self.vFuncFxd_next = self.solution_next.vFuncFxd
-
-        # If the distributions are NOT independent...
-        # Unpack the shock distribution
-        self.TranShks_next = self.ShockDstn.atoms[1]
-        # Flag for whether the natural borrowing constraint is zero
-        self.zero_bound = np.min(self.TranShks_next) == 0.0
-
-    def prepare_to_calc_EndOfPrdvP(self):
-        """
-        Prepare to calculate end-of-period marginal values by creating an array
-        of market resources that the agent could have next period, considering
-        the grid of end-of-period assets and the distribution of shocks he might
-        experience next period.
-        """
-
-        # Make tiled arrays to calculate future realizations of mNrm and Share; dimension order: mNrm, Share, shock
-        if self.zero_bound:
-            self.aNrmGrid = self.aXtraGrid
-        else:
-            # Add an asset point at exactly zero
-            self.aNrmGrid = np.insert(self.aXtraGrid, 0, 0.0)
-
-        self.aNrmCount = self.aNrmGrid.size
-        self.ShareCount = self.ShareGrid.size
-
-        self.aNrm_tiled, self.Share_tiled = np.meshgrid(
-            self.aNrmGrid, self.ShareGrid, indexing="ij"
-        )
-
-    def r_port(self, shocks, share):
-        """
-        Calculate future realizations of market resources
-        """
-
-        return (1.0 - share) * self.Rfree + share * shocks["Risky"]
-
-    def m_nrm_next(self, shocks, a_nrm, r_port):
-        """
-        Calculate future realizations of market resources
-        """
-
-        return (
-            r_port * a_nrm / (shocks["PermShk"] * self.PermGroFac) + shocks["TranShk"]
-        )
-
-    def calc_EndOfPrdvP(self):
-        """
-        Calculate end-of-period marginal value of assets and shares at each point
-        in aNrm and ShareGrid. Does so by taking expectation of next period marginal
-        values across income and risky return shocks.
-        """
-
-        def dvdm(m_nrm_next, shares):
-            """
-            Evaluate realizations of marginal value of market resources next period
-            """
-
-            dvdmAdj_next = self.vPfuncAdj_next(m_nrm_next)
-            if self.AdjustPrb < 1.0:
-                dvdmFxd_next = self.dvdmFuncFxd_next(m_nrm_next, shares)
-                # Combine by adjustment probability
-                dvdm_next = (
-                    self.AdjustPrb * dvdmAdj_next
-                    + (1.0 - self.AdjustPrb) * dvdmFxd_next
-                )
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                dvdm_next = dvdmAdj_next
-
-            return dvdm_next
-
-        def dvds(m_nrm_next, shares):
-            """
-            Evaluate realizations of marginal value of risky share next period
-            """
-
-            # No marginal value of Share if it's a free choice!
-            dvdsAdj_next = np.zeros_like(m_nrm_next)
-            if self.AdjustPrb < 1.0:
-                dvdsFxd_next = self.dvdsFuncFxd_next(m_nrm_next, shares)
-                # Combine by adjustment probability
-                dvds_next = (
-                    self.AdjustPrb * dvdsAdj_next
-                    + (1.0 - self.AdjustPrb) * dvdsFxd_next
-                )
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                dvds_next = dvdsAdj_next
-
-            return dvds_next
-
-        def EndOfPrddvda_dists(shocks, a_nrm, shares):
-            r_port = self.r_port(shocks, shares)
-            m_nrm_next = self.m_nrm_next(shocks, a_nrm, r_port)
-
-            # Expand shares to the shape of m so that operations can be vectorized
-            shares_expanded = shares + np.zeros_like(m_nrm_next)
-
-            return (
-                r_port
-                * self.uP(shocks["PermShk"] * self.PermGroFac)
-                * dvdm(m_nrm_next, shares_expanded)
-            )
-
-        def EndOfPrddvds_dist(shocks, a_nrm, shares):
-            Rxs = shocks["Risky"] - self.Rfree
-            r_port = self.r_port(shocks, shares)
-            m_nrm_next = self.m_nrm_next(shocks, a_nrm, r_port)
-
-            return Rxs * a_nrm * self.uP(shocks["PermShk"] * self.PermGroFac) * dvdm(
-                m_nrm_next, shares
-            ) + (shocks["PermShk"] * self.PermGroFac) ** (1.0 - self.CRRA) * dvds(
-                m_nrm_next, shares
-            )
-
-        # Calculate end-of-period marginal value of assets by taking expectations
-        self.EndOfPrddvda = (
-            self.DiscFac
-            * self.LivPrb
-            * self.ShockDstn.expected(
-                EndOfPrddvda_dists, self.aNrm_tiled, self.Share_tiled
-            )
-        )
-
-        self.EndOfPrddvdaNvrs = self.uPinv(self.EndOfPrddvda)
-
-        # Calculate end-of-period marginal value of risky portfolio share by taking expectations
-        self.EndOfPrddvds = (
-            self.DiscFac
-            * self.LivPrb
-            * self.ShockDstn.expected(
-                EndOfPrddvds_dist, self.aNrm_tiled, self.Share_tiled
-            )
-        )
-
-    def make_EndOfPrdvFunc(self):
-        """
-        Construct the end-of-period value function for this period, storing it
-        as an attribute of self for use by other methods.
-        """
-
-        def v_dist(shocks, a_nrm, shares):
-            r_port = self.r_port(shocks, shares)
-            m_nrm_next = self.m_nrm_next(shocks, a_nrm, r_port)
-
-            vAdj_next = self.vFuncAdj_next(m_nrm_next)
-            if self.AdjustPrb < 1.0:
-                vFxd_next = self.vFuncFxd_next(m_nrm_next, shares)
-                v_next = self.AdjustPrb * vAdj_next + (1.0 - self.AdjustPrb) * vFxd_next
-            else:  # Don't bother evaluating if there's no chance that portfolio share is fixed
-                v_next = vAdj_next
-
-            return (shocks["PermShk"] * self.PermGroFac) ** (1.0 - self.CRRA) * v_next
-
-        self.EndOfPrdv = (
-            self.DiscFac
-            * self.LivPrb
-            * self.ShockDstn.expected(v_dist, self.aNrm_tiled, self.Share_tiled)
-        )
-
-        self.EndOfPrdvNvrs = self.uinv(self.EndOfPrdv)
-
-    def solve(self):
-        """
-        Solve the one period problem for a portfolio-choice consumer.
-
-        Returns
-        -------
-        solution_now : PortfolioSolution
-        The solution to the single period consumption-saving with portfolio choice
-        problem.  Includes two consumption and risky share functions: one for when
-        the agent can adjust his portfolio share (Adj) and when he can't (Fxd).
-        """
-
-        # Make arrays of end-of-period assets and end-of-period marginal value
-        self.prepare_to_calc_EndOfPrdvP()
-        self.calc_EndOfPrdvP()
-
-        if self.DiscreteShareBool:
-            self.make_EndOfPrdvFunc()
-            ConsPortfolioDiscreteSolver.optimize_share(self)
-
-            # Construct a basic solution for this period
-            self.make_basic_solution()
-            ConsPortfolioDiscreteSolver.make_ShareFuncAdj(self)
-            self.make_vFunc()
-        else:
-            # Construct a basic solution for this period
-            ConsPortfolioSolver.optimize_share(self)
-            self.make_basic_solution()
-            ConsPortfolioSolver.make_ShareFuncAdj(self)
-
-            # Add the value function if requested
-            if self.vFuncBool:
-                self.add_vFunc()
-            else:  # If vFuncBool is False, fill in dummy values
-                self.vFuncAdj_now = NullFunc()
-                self.vFuncFxd_now = NullFunc()
-
-        self.add_save_points()
-
-        self.make_porfolio_solution()
-
-        return self.solution
-
-
-class ConsSequentialPortfolioSolver(ConsPortfolioSolver):
-    def add_SequentialShareFuncAdj(self, solution):
-        """
-        Construct the risky share function as a function of savings when the agent can adjust.
-        """
-
-        if self.zero_bound:
-            Share_lower_bound = self.ShareLimit
-            aNrm_temp = np.insert(self.aNrmGrid, 0, 0.0)
-            Share_now = np.insert(self.Share_now, 0, Share_lower_bound)
-        else:
-            aNrm_temp = self.aNrmGrid  # already includes 0.0
-            Share_now = self.Share_now
-
-        self.SequentialShareFuncAdj_now = LinearInterp(
-            aNrm_temp,
-            Share_now,
-            intercept_limit=self.ShareLimit,
-            slope_limit=0.0,
-        )
-
-        solution.SequentialShareFuncAdj = self.SequentialShareFuncAdj_now
-
-        return solution
-
-    def solve(self):
-        solution = ConsPortfolioSolver.solve(self)
-
-        solution = self.add_SequentialShareFuncAdj(solution)
-
-        return solution
 
 
 # Make a dictionary to specify a portfolio choice consumer type
